Changed Player Movement to new Vector Class Based Movement. Added Diagonal Movement to the player using a directional vector. Added vector normalization function.

+++ 2D-Game/src/object/Eye_Obj.java
@@ -28,8 +28,8 @@ public class Eye_Obj extends SuperObject{
 		
 		solidArea.width = gp.tileSize;
 		solidArea.height = gp.tileSize;
-		int screenX = worldPos.x - gp.player.worldPos.x + gp.player.screenX;
-		int screenY = worldPos.y - gp.player.worldPos.y + gp.player.screenY;
+		int screenX = (int) (worldPos.x - gp.player.worldPos.x + gp.player.screenX);
+		int screenY = (int) (worldPos.y - gp.player.worldPos.y + gp.player.screenY);
 		if(worldPos.x < gp.player.worldPos.x + gp.player.screenX + gp.tileSize
 			&& worldPos.x > gp.player.worldPos.x - gp.player.screenX - gp.tileSize
 			&& worldPos.y > gp.player.worldPos.y - gp.player.screenY - gp.tileSize
